+++ internal/framework/search_test.go
@@ -8,12 +8,13 @@ import (
 
 	"github.com/Masterminds/semver/v3"
 	"github.com/saucelabs/saucectl/internal/node"
+	"github.com/saucelabs/saucectl/internal/runtime"
 )
 
 type MockMetadataService struct {
 	MockFrameworks func(ctx context.Context) ([]string, error)
 	MockVersions   func(ctx context.Context, frameworkName string) ([]Metadata, error)
-	MockRuntimes   func(ctx context.Context) ([]Runtime, error)
+	MockRuntimes   func(ctx context.Context) ([]runtime.Runtime, error)
 }
 
 func (m *MockMetadataService) Frameworks(ctx context.Context) ([]string, error) {
@@ -32,7 +33,7 @@ func (m *MockMetadataService) Versions(ctx context.Context, frameworkName string
 	return nil, nil
 }
 
-func (m *MockMetadataService) Runtimes(ctx context.Context) ([]Runtime, error) {
+func (m *MockMetadataService) Runtimes(ctx context.Context) ([]runtime.Runtime, error) {
 	if m.MockRuntimes != nil {
 		return m.MockRuntimes(ctx)
 	}
