+++ internal/mocks/frameworks.go
@@ -4,13 +4,14 @@ import (
 	"context"
 
 	"github.com/saucelabs/saucectl/internal/framework"
+	"github.com/saucelabs/saucectl/internal/runtime"
 )
 
 // FakeFrameworkInfoReader is a mock for the interface framework.MetadataService.
 type FakeFrameworkInfoReader struct {
 	FrameworksFn func(ctx context.Context) ([]string, error)
 	VersionsFn   func(ctx context.Context, frameworkName string) ([]framework.Metadata, error)
-	RuntimeFn    func(ctx context.Context) ([]framework.Runtime, error)
+	RuntimeFn    func(ctx context.Context) ([]runtime.Runtime, error)
 }
 
 // Frameworks is a wrapper around FrameworksFn.
@@ -24,6 +25,6 @@ func (fir *FakeFrameworkInfoReader) Versions(ctx context.Context, frameworkName 
 }
 
 // Runtimes is a wrapper around RuntimesFn.
-func (fir *FakeFrameworkInfoReader) Runtimes(ctx context.Context) ([]framework.Runtime, error) {
+func (fir *FakeFrameworkInfoReader) Runtimes(ctx context.Context) ([]runtime.Runtime, error) {
 	return fir.RuntimeFn(ctx)
 }
