+++ internal/saucecloud/cucumber.go
@@ -10,6 +10,7 @@ import (
 	"github.com/saucelabs/saucectl/internal/framework"
 	"github.com/saucelabs/saucectl/internal/msg"
 	"github.com/saucelabs/saucectl/internal/playwright"
+	"github.com/saucelabs/saucectl/internal/runtime"
 
 	"github.com/saucelabs/saucectl/internal/job"
 )
@@ -39,14 +40,14 @@ func (r *CucumberRunner) RunProject() (int, error) {
 		if err != nil {
 			return 1, err
 		}
-		runtime, err := framework.SelectNode(runtimes, r.Project.NodeVersion)
+		rt, err := runtime.SelectNode(runtimes, r.Project.NodeVersion)
 		if err != nil {
 			return 1, err
 		}
-		if err := framework.ValidateRuntime(runtime); err != nil {
+		if err := rt.Validate(); err != nil {
 			return 1, err
 		}
-		r.Project.NodeVersion = runtime.RuntimeVersion
+		r.Project.NodeVersion = rt.RuntimeVersion
 	}
 	// If the runner doesn't support the global node, set nodeVersion to empty.
 	if !m.SupportGlobalNode() {
